Refactor credentials management and clients
