+++ integrations/messenger/src/actions/proactive-conversation.ts
@@ -1,5 +1,5 @@
 import { RuntimeError } from '@botpress/sdk'
-import { getMessengerClient } from '../misc/utils'
+import { create as createMessengerClient } from '../misc/messenger-client'
 import * as bp from '.botpress'
 
 const getOrCreateConversation: bp.IntegrationProps['actions']['getOrCreateConversation'] = async ({
@@ -12,7 +12,7 @@ const getOrCreateConversation: bp.IntegrationProps['actions']['getOrCreateConver
     throw new RuntimeError('User ID is required')
   }
 
-  const messengerClient = await getMessengerClient(client, ctx)
+  const messengerClient = await createMessengerClient(client, ctx)
   const profile = await messengerClient.getUserProfile(userId)
 
   const { conversation } = await client.getOrCreateConversation({
