+++ integrations/messenger/src/actions/proactive-user.ts
@@ -1,5 +1,5 @@
 import { RuntimeError } from '@botpress/sdk'
-import { getMessengerClient } from '../misc/utils'
+import { create as createMessengerClient } from '../misc/messenger-client'
 import * as bp from '.botpress'
 
 const getOrCreateUser: bp.IntegrationProps['actions']['getOrCreateUser'] = async ({ client, ctx, input }) => {
@@ -8,7 +8,7 @@ const getOrCreateUser: bp.IntegrationProps['actions']['getOrCreateUser'] = async
     throw new RuntimeError('User ID is required')
   }
 
-  const messengerClient = await getMessengerClient(client, ctx)
+  const messengerClient = await createMessengerClient(client, ctx)
   const profile = await messengerClient.getUserProfile(userId)
 
   const { user } = await client.getOrCreateUser({ tags: { id: profile.id } })
