+++ integrations/messenger/src/webhook/handlers/oauth/wizard.ts
@@ -1,5 +1,5 @@
 import * as oauthWizard from '@botpress/common/src/oauth-wizard'
-import { MetaClient } from '../../../misc/client'
+import { MetaClient } from '../../../misc/meta-client'
 import * as bp from '.botpress'
 import { Oauth as OAuthState } from '.botpress/implementation/typings/states/oauth'
 
@@ -90,7 +90,7 @@ const _oauthCallbackHandler: WizardHandler = async ({ responses, query, client, 
 
 const _selectPageHandler: WizardHandler = async ({ responses, client, ctx, logger }) => {
   const metaClient = new MetaClient(logger)
-  const { accessToken } = await _getCredentials(client, ctx)
+  const { accessToken } = await _getOAuthCredentials(client, ctx)
   if (!accessToken) {
     return responses.endWizard({
       success: false,
@@ -113,7 +113,7 @@ const _selectPageHandler: WizardHandler = async ({ responses, client, ctx, logge
 
 const _setupHandler: WizardHandler = async ({ responses, client, ctx, logger, selectedChoice }) => {
   const metaClient = new MetaClient(logger)
-  const { accessToken } = await _getCredentials(client, ctx)
+  const { accessToken } = await _getOAuthCredentials(client, ctx)
   if (!accessToken) {
     return responses.endWizard({
       success: false,
@@ -182,7 +182,7 @@ const _getOAuthRedirectUri = (ctx?: bp.Context) => oauthWizard.getWizardStepUrl(
 
 // client.patchState is not working correctly
 const _patchCredentials = async (client: bp.Client, ctx: bp.Context, newState: Partial<OAuthState['payload']>) => {
-  const currentState = await _getCredentials(client, ctx).catch(() => ({}))
+  const currentState = await _getOAuthCredentials(client, ctx).catch(() => ({}))
 
   await client.setState({
     type: 'integration',
@@ -195,7 +195,7 @@ const _patchCredentials = async (client: bp.Client, ctx: bp.Context, newState: P
   })
 }
 
-const _getCredentials = async (client: bp.Client, ctx: bp.Context) => {
+const _getOAuthCredentials = async (client: bp.Client, ctx: bp.Context) => {
   return await client
     .getState({ type: 'integration', name: 'oauth', id: ctx.integrationId })
     .then((result) => result.state.payload)
